refactor(view-engine): extract AnalysisSortEditor
